Encapsulate NLP Virtual Graph logic
